Adicionando metodo delete e update em pessoa

+++ src/main/java/br/ufrn/imd/utravel/repository/GenericRepository.java
@@ -15,5 +15,5 @@ public interface GenericRepository <T extends AbstractModel> {
 
     T update(T modelo);
 
-    String delete(Long id);
+    String delete(Integer id);
 }
